Library POC
Got libary working with the double arry.

Got to clean up the very confusing naming scheme.
This will be fixed as the cleanup comes through.

Then got to fix it up to be a bit more generic.
Functionality wise - got to add loading from disk and genering libaries in the backend.